+++ cli/src/cli_util.rs
@@ -81,6 +81,7 @@ use jj_lib::fileset::FilesetExpression;
 use jj_lib::fileset::FilesetParseContext;
 use jj_lib::gitignore::GitIgnoreError;
 use jj_lib::gitignore::GitIgnoreFile;
+use jj_lib::gitignore::GitIgnoreFollowSymlinks;
 use jj_lib::id_prefix::IdPrefixContext;
 use jj_lib::lock::FileLock;
 use jj_lib::matchers::Matcher;
@@ -1699,16 +1700,25 @@ to the current parents may contain changes from multiple commits.
         if let Ok(git_backend) = jj_lib::git::get_git_backend(self.repo().store()) {
             let git_repo = git_backend.git_repo();
             if let Some(excludes_file_path) = get_excludes_file_path(&git_repo.config_snapshot()) {
-                git_ignores = git_ignores.chain_with_file(RepoPath::root(), excludes_file_path)?;
+                git_ignores = git_ignores.chain_with_file(
+                    RepoPath::root(),
+                    &excludes_file_path,
+                    GitIgnoreFollowSymlinks::Follow,
+                )?;
             }
             git_ignores = git_ignores.chain_with_file(
                 RepoPath::root(),
-                git_backend.git_repo_path().join("info").join("exclude"),
+                &git_backend.git_repo_path().join("info").join("exclude"),
+                GitIgnoreFollowSymlinks::Follow,
             )?;
         } else if let Ok(git_config) = gix::config::File::from_globals()
             && let Some(excludes_file_path) = get_excludes_file_path(&git_config)
         {
-            git_ignores = git_ignores.chain_with_file(RepoPath::root(), excludes_file_path)?;
+            git_ignores = git_ignores.chain_with_file(
+                RepoPath::root(),
+                &excludes_file_path,
+                GitIgnoreFollowSymlinks::Follow,
+            )?;
         }
         Ok(git_ignores)
     }